18 improve marker management

+++ src/assets/classes/MarkerList.js
@@ -1,6 +1,9 @@
 import { Marker } from './Marker.js'
+import { Modal } from './Modal.js'
 import { LoopManager } from './LoopManager.js'
 import { Song } from './Song.js'
+import { Notification } from './Notification.js'
+import { formattedTimeToSeconds, secondsToFormattedTime } from '../utils.js'
 
 export class MarkerList {
 	/**
@@ -113,9 +116,9 @@ export class MarkerList {
 			const loopCheckbox = this.getLoopCheckbox(marker)
 			item.appendChild(loopCheckbox)
 			item.appendChild(this.getLoopProxy(loopCheckbox, marker))
-			item.appendChild(this.getDeleteButton(marker))
 			item.appendChild(this.getButton(marker))
 			item.appendChild(this.getInput(marker))
+			item.appendChild(this.getEditMarkerButton(marker))
 
 			// Append the item to the list
 			list.appendChild(item)
@@ -169,6 +172,164 @@ export class MarkerList {
 	}
 
 	/**
+	 * Returns an edit marker button
+	 * @param {Marker} marker - A Marker instance
+	 * @returns {HTMLButtonElement} - A button element
+	*/
+	getEditMarkerButton(marker) {
+		const button = document.createElement('button')
+		button.textContent = 'Edit Marker'
+		button.addEventListener('click', () => {
+			const modalHeader = document.createElement('h2')
+			modalHeader.textContent = marker.title
+
+			// Append buttons to modal header and get edit form content
+			modalHeader.appendChild(this.getEditTitleButton(marker, modalHeader))
+			modalHeader.appendChild(this.getDeleteButton(marker))
+			const modalContent = this.getEditMarkerForm(marker)
+
+			// Open modal
+			this.activeModal = new Modal(modalHeader, modalContent, { useForm: true })
+		})
+		return button
+	}
+
+	/**
+	 * Returns an edit title button for the marker
+	 * @param {Marker} marker - A Marker instance
+	 * @param {HTMLHeadingElement} modalHeader - A heading element
+	 * @returns {HTMLButtonElement} - A button element
+	*/
+	getEditTitleButton(marker, modalHeader) {
+		const button = document.createElement('button')
+		button.classList.add('edit-asset-title')
+		button.innerHTML = '&#9998;'
+		button.ariaLabel = 'Edit title'
+		button.addEventListener('click', () => {
+			modalHeader.innerHTML = ''
+			const titleInput = document.createElement('input')
+			titleInput.type = 'text'
+			titleInput.value = marker.getTitle()
+
+			const saveButton = document.createElement('button')
+			saveButton.textContent = 'Save'
+			saveButton.addEventListener('click', (e) => {
+				e.preventDefault()
+				marker.setTitle(titleInput.value)
+				marker.song.bandbook.syncManager.updateMarkerTitle(marker, titleInput.value)
+				marker.song.bandbook.refresh()
+
+				// Update modal header
+				modalHeader.textContent = marker.title
+				modalHeader.appendChild(this.getEditTitleButton(marker, modalHeader))
+				modalHeader.appendChild(this.getDeleteButton(marker))
+			})
+
+			modalHeader.appendChild(titleInput)
+			modalHeader.appendChild(saveButton)
+		})
+		return button
+	}
+		
+
+	/**
+	 * Returns an edit form for the marker
+	 * @returns {HTMLDivElement} - A div wrapper around the form element
+	*/
+	getEditMarkerForm(marker) {
+		const hiddenInputChange = (e) => {
+			const time = e.target.value
+			marker.setTime(time)
+			marker.song.bandbook.refresh()
+		}
+
+		const div = document.createElement('div')
+		div.classList.add('edit-asset')
+
+		// time input
+		const timeInput = document.createElement('input')
+		timeInput.type = 'number'
+		timeInput.step = "1"
+		timeInput.value = Math.floor(marker.time)
+		timeInput.hidden = true
+		timeInput.addEventListener('change', hiddenInputChange)
+		div.appendChild(timeInput)
+
+		// time proxy
+		const timeProxyWrapper = document.createElement('div')
+		timeProxyWrapper.classList.add('time-proxy-wrapper')
+		const timeProxy = document.createElement('input')
+		timeProxy.type = 'text'
+		// Smart HH:MM:SS format
+		timeProxy.pattern = "^(?:(?:1[0-1]|[1-9]):)?(?:[0-5][0-9]:)?[0-5][0-9]$"
+		timeProxy.value = marker.getFormattedTime()
+		timeProxy.addEventListener('change', () => {
+			const previousTime = marker.getFormattedTime()
+
+			// Validate
+			if (!timeProxy.value.match(timeProxy.pattern)) {
+				const error = new Notification('Time must be in valid format: SS, MM:SS, or HH:MM:SS', 'error', true, 5000, true)
+				timeProxyWrapper.insertAdjacentElement('afterend', error.element)
+				timeProxy.value = previousTime
+				return
+			}
+			
+			const time = formattedTimeToSeconds(timeProxy.value)
+
+			if (time < 0 || time > marker.song.getDuration()) {
+				const error = new Notification('That is not a valid time for this song', 'error', true, 5000, true)
+				timeProxyWrapper.insertAdjacentElement('afterend', error.element)
+				timeProxy.value = previousTime
+				return
+			}
+
+			hiddenInputChange({ target: { value: time } })
+		})
+
+		const upOneSecond = document.createElement('button')
+		upOneSecond.innerHTML = '&#9650;'
+		upOneSecond.addEventListener('click', (e) => {
+			e.preventDefault()
+			const time = formattedTimeToSeconds(timeProxy.value) + 1
+			hiddenInputChange({ target: { value: time } })
+			timeProxy.value = secondsToFormattedTime(time)
+		})
+		const downOneSecond = document.createElement('button')
+		downOneSecond.innerHTML = '&#9660;'
+		downOneSecond.addEventListener('click', (e) => {
+			e.preventDefault()
+			const time = formattedTimeToSeconds(timeProxy.value) - 1
+			hiddenInputChange({ target: { value: time } })
+			timeProxy.value = secondsToFormattedTime(time)
+		})
+		
+		timeProxyWrapper.appendChild(timeProxy)
+		timeProxyWrapper.appendChild(upOneSecond)
+		timeProxyWrapper.appendChild(downOneSecond)
+		div.appendChild(timeProxyWrapper)
+
+		// notes input
+		const notesLabel = document.createElement('label')
+		notesLabel.htmlFor = 'marker-notes'
+		const notesSpan = document.createElement('span')
+		notesSpan.textContent = 'Notes'
+		const notesInput = document.createElement('textarea')
+		notesInput.value = marker.getNotes()
+		notesInput.id = 'marker-notes'
+		notesInput.name = 'marker-notes'
+		notesInput.rows = 4
+		notesInput.addEventListener('change', () => {
+			marker.setNotes(notesInput.value)
+		})
+		notesLabel.appendChild(notesSpan)
+		notesLabel.appendChild(notesInput)
+
+		div.appendChild(notesLabel)
+
+		return div
+	}
+
+	/**
 	 * Returns a delete button for the marker
 	 * @param {Marker} marker - A Marker instance
 	 * @returns {HTMLButtonElement} - A button element
@@ -177,9 +338,14 @@ export class MarkerList {
 		const deleteButton = document.createElement('button')
 		deleteButton.textContent = 'Delete'
 		deleteButton.addEventListener('click', () => {
-			this.removeMarker(marker)
-			this.renderMarkersList()
-			this.song.bandbook.syncManager.deleteMarker(marker)
+			if (confirm(`Are you sure you want to delete ${marker.title}?`)) {
+				this.removeMarker(marker)
+				this.song.bandbook.syncManager.deleteMarker(marker)
+				this.song.bandbook.refresh()
+				document.dispatchEvent(new KeyboardEvent('keydown', { key: 'Escape' }))
+			} else {
+				return
+			}
 		})
 		return deleteButton
 	}
